+++ modules/core/dependency/python-ihm/ihm/citations.py
@@ -15,8 +15,9 @@ imp = ihm.Citation(
     title='Putting the pieces together: integrative modeling platform '
           'software for structure determination of macromolecular assemblies',
     journal='PLoS Biol', volume=10, page_range='e1001244', year=2012,
-    authors=['Russel D', 'Lasker K', 'Webb B', 'Velázquez-Muriel J', 'Tjioe E',
-             'Schneidman-Duhovny D', 'Peterson B', 'Sali A'],
+    authors=['Russel, D.', 'Lasker, K.', 'Webb, B.', 'Velázquez-Muriel, J.',
+             'Tjioe, E.', 'Schneidman-Duhovny, D.', 'Peterson, B.',
+             'Sali, A.'],
     doi='10.1371/journal.pbio.1001244')
 
 
@@ -24,8 +25,9 @@ pmi = ihm.Citation(
     pmid='31396911',
     title='Modeling Biological Complexes Using Integrative Modeling Platform.',
     journal='Methods Mol Biol', volume=2022, page_range=(353, 377), year=2019,
-    authors=['Saltzberg D', 'Greenberg CH', 'Viswanath S', 'Chemmama I',
-             'Webb B', 'Pellarin R', 'Echeverria I', 'Sali A'],
+    authors=['Saltzberg, D.', 'Greenberg, C.H.', 'Viswanath, S.',
+             'Chemmama, I.', 'Webb, B.', 'Pellarin, R.', 'Echeverria, I.',
+             'Sali, A.'],
     doi='10.1007/978-1-4939-9608-7_15')
 
 
@@ -34,7 +36,7 @@ modeller = ihm.Citation(
     title='Comparative protein modelling by satisfaction of '
           'spatial restraints.',
     journal='J Mol Biol', volume=234, page_range=(779, 815), year=1993,
-    authors=['Sali A', 'Blundell TL'], doi='10.1006/jmbi.1993.1626')
+    authors=['Sali, A.', 'Blundell, T.L.'], doi='10.1006/jmbi.1993.1626')
 
 
 psipred = ihm.Citation(
@@ -42,7 +44,7 @@ psipred = ihm.Citation(
     title='Protein secondary structure prediction based on position-specific '
           'scoring matrices.',
     journal='J Mol Biol', volume=292, page_range=(195, 202), year=1999,
-    authors=['Jones DT'], doi='10.1006/jmbi.1999.3091')
+    authors=['Jones, D.T.'], doi='10.1006/jmbi.1999.3091')
 
 
 disopred = ihm.Citation(
@@ -50,7 +52,7 @@ disopred = ihm.Citation(
     title='DISOPRED3: precise disordered region predictions with annotated '
           'protein-binding activity.',
     journal='Bioinformatics', volume=31, page_range=(857, 863), year=2015,
-    authors=['Jones DT', 'Cozzetto D'], doi='10.1093/bioinformatics/btu744')
+    authors=['Jones, D.T.', 'Cozzetto D'], doi='10.1093/bioinformatics/btu744')
 
 
 hhpred = ihm.Citation(
@@ -58,7 +60,7 @@ hhpred = ihm.Citation(
     title='The HHpred interactive server for protein homology detection '
           'and structure prediction.',
     journal='Nucleic Acids Res', volume=33, page_range=('W244', 'W248'),
-    year=2005, authors=['Söding J', 'Biegert A', 'Lupas AN'],
+    year=2005, authors=['Söding, J.', 'Biegert, A.', 'Lupas, A.N.'],
     doi='10.1093/nar/gki408')
 
 
@@ -67,7 +69,7 @@ relion = ihm.Citation(
     title='RELION: implementation of a Bayesian approach to cryo-EM '
           'structure determination.',
     journal='J Struct Biol', volume=180, page_range=(519, 530), year=2012,
-    authors=['Scheres SH'], doi='10.1016/j.jsb.2012.09.006')
+    authors=['Scheres, S.H.'], doi='10.1016/j.jsb.2012.09.006')
 
 
 phyre2 = ihm.Citation(
@@ -75,7 +77,8 @@ phyre2 = ihm.Citation(
     title='The Phyre2 web portal for protein modeling, prediction '
           'and analysis.',
     journal='Nat Protoc', volume=10, page_range=('845', '858'), year=2015,
-    authors=['Kelley LA', 'Mezulis S', 'Yates CM', 'Wass MN', 'Sternberg MJ'],
+    authors=['Kelley, L.A.', 'Mezulis, S.', 'Yates, C.M.', 'Wass, M.N.',
+             'Sternberg, M.J.'],
     doi='10.1038/nprot.2015.053')
 
 
@@ -85,9 +88,9 @@ swiss_model = ihm.Citation(
           'and complexes.',
     journal='Nucleic Acids Res', volume=46, page_range=('W296', 'W303'),
     year=2018,
-    authors=['Waterhouse A', 'Bertoni M', 'Bienert S', 'Studer G',
-             'Tauriello G', 'Gumienny R', 'Heer FT', 'de Beer TAP',
-             'Rempfer C', 'Bordoli L', 'Lepore R', 'Schwede T'],
+    authors=['Waterhouse, A.', 'Bertoni, M.', 'Bienert, S.', 'Studer, G.',
+             'Tauriello, G.', 'Gumienny, R.', 'Heer, F.T.', 'de Beer, T.A.P.',
+             'Rempfer, C.', 'Bordoli, L.', 'Lepore, R.', 'Schwede, T.'],
     doi='10.1093/nar/gky427')
 
 
@@ -95,12 +98,22 @@ alphafold2 = ihm.Citation(
     pmid='34265844',
     title='Highly accurate protein structure prediction with AlphaFold.',
     journal='Nature', volume=596, page_range=(583, 589), year=2021,
-    authors=['Jumper J', 'Evans R', 'Pritzel A', 'Green T', 'Figurnov M',
-             'Ronneberger O', 'Tunyasuvunakool K', 'Bates R', 'Zidek A',
-             'Potapenko A', 'Bridgland A', 'Meyer C', 'Kohl SAA', 'Ballard AJ',
-             'Cowie A', 'Romera-Paredes B', 'Nikolov S', 'Jain R', 'Adler J',
-             'Back T', 'Petersen S', 'Reiman D', 'Clancy E', 'Zielinski M',
-             'Steinegger M', 'Pacholska M', 'Berghammer T', 'Bodenstein S',
-             'Silver D', 'Vinyals O', 'Senior AW', 'Kavukcuoglu K', 'Kohli P',
-             'Hassabis D'],
+    authors=['Jumper, J.', 'Evans, R.', 'Pritzel, A.', 'Green, T.',
+             'Figurnov, M.', 'Ronneberger, O.', 'Tunyasuvunakool, K.',
+             'Bates, R.', 'Zidek, A.', 'Potapenko, A.', 'Bridgland, A.',
+             'Meyer, C.', 'Kohl, S.A.A.', 'Ballard, A.J.', 'Cowie, A.',
+             'Romera-Paredes, B.', 'Nikolov, S.', 'Jain, R.', 'Adler, J.',
+             'Back, T.', 'Petersen, S.', 'Reiman, D.', 'Clancy, E.',
+             'Zielinski, M.', 'Steinegger, M.', 'Pacholska, M.',
+             'Berghammer, T.', 'Bodenstein, S.', 'Silver, D.', 'Vinyals, O.',
+             'Senior, A.W.', 'Kavukcuoglu, K.', 'Kohli, P.', 'Hassabis, D.'],
     doi='10.1038/s41586-021-03819-2')
+
+
+colabfold = ihm.Citation(
+    pmid='35637307',
+    title='ColabFold: making protein folding accessible to all.',
+    journal='Nature Methods', volume=19, page_range=(679, 682), year=2022,
+    authors=['Mirdita, M.', 'Schuetze, K.', 'Moriwaki, Y.', 'Heo, L.',
+             'Ovchinnikov, S.', 'Steinegger, M.'],
+    doi='10.1038/s41592-022-01488-1')